# Conflicts:
#	Server/MarketData/insert_quotes.py
#	Server/StockTKL/README.md
#	Server/StockTKL/src/main/java/com/example/stocktkl/controller/OrderController.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/model/Quote.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/model/enum_class/EOrderType.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/payload/request/OrderRequest.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/repository/OrderRepository.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/service/IOrderService.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/service/OrderService.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/service/impl/OrderService.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/service/impl/OrderServiceImpl.java
#	Server/StockTKL/src/main/java/com/example/stocktkl/trading/TradingService.java
#	Server/StockTKL/src/main/resources/application.properties

+++ Server/StockTKL/src/main/java/com/example/stocktkl/payload/request/OrderRequest.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,5 +17,5 @@ public class OrderRequest {
     private EOrderType orderType;
     private EOrderDirection direction;
     private int quantity;
-    private int price = -1;
+    private BigDecimal price = BigDecimal.valueOf(-1);
 }
